requirements.txt
Flask==3.1.3
google-genai==1.47.0
gunicorn==23.0.0
pdfrw==0.4
Pillow==11.3.0
pypdf==6.9.2
python-dotenv==1.2.1
reportlab==4.4.10
